+++ src/main/java/ru/vyarus/dropwizard/guice/module/context/unique/item/UniqueGuiceyBundle.java
@@ -1,5 +1,6 @@
 package ru.vyarus.dropwizard.guice.module.context.unique.item;
 
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import ru.vyarus.dropwizard.guice.module.installer.bundle.GuiceyBundle;
 
 /**
@@ -7,23 +8,27 @@ import ru.vyarus.dropwizard.guice.module.installer.bundle.GuiceyBundle;
  * Note that class only properly implements equals method so guicey deduplication mechanism could filter other
  * instances. It is not required to use this class to grant bundle uniqueness - you may directly implement equals
  * method in your bundle.
+ * <p>
+ * Classed are compared by name to properly detect classes from different class loaders.
  *
  * @author Vyacheslav Rusakov
  * @see ru.vyarus.dropwizard.guice.GuiceBundle.Builder#duplicateConfigDetector(
  *ru.vyarus.dropwizard.guice.module.context.unique.DuplicateConfigDetector)
  * @since 13.07.2019
  */
+@SuppressFBWarnings("EQ_COMPARING_CLASS_NAMES")
 public abstract class UniqueGuiceyBundle implements GuiceyBundle {
 
     @Override
     public boolean equals(final Object obj) {
         // only one bundle instance allowed
-        return obj != null && getClass().equals(obj.getClass());
+        // intentionally check by class name to also detect instances from different class loaders
+        return obj != null && getClass().getName().equals(obj.getClass().getName());
     }
 
     @Override
     public int hashCode() {
         // for data structures relying on hash first, all equal instances must have the same hash to avoid side effects
-        return getClass().hashCode();
+        return getClass().getName().hashCode();
     }
 }