+++ src/main/java/ru/vyarus/dropwizard/guice/module/context/unique/item/UniqueModule.java
@@ -1,12 +1,15 @@
 package ru.vyarus.dropwizard.guice.module.context.unique.item;
 
 import com.google.inject.AbstractModule;
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
 /**
  * Base class for unique modules: only one module instance must be accepted and all others considered duplicate.
  * Note that class only properly implements equals method so guicey deduplication mechanism could filter other
  * instances. It is not required to use this class to grant module uniqueness - you may directly implement equals
  * method in your module.
+ * <p>
+ * Classed are compared by name to properly detect classes from different class loaders.
  * <p>
  * Also, note that guice silently ignores duplicate bindings and so in most cases it would be able to
  * handle duplicate modules properly. This class may be used for cases when bindings registered by module
@@ -17,17 +20,19 @@ import com.google.inject.AbstractModule;
  *ru.vyarus.dropwizard.guice.module.context.unique.DuplicateConfigDetector)
  * @since 13.07.2019
  */
+@SuppressFBWarnings("EQ_COMPARING_CLASS_NAMES")
 public abstract class UniqueModule extends AbstractModule {
 
     @Override
     public boolean equals(final Object obj) {
-        // only one debug module instance allowed
-        return obj != null && getClass().equals(obj.getClass());
+        // only one module instance allowed
+        // intentionally check by class name to also detect instances from different class loaders
+        return obj != null && getClass().getName().equals(obj.getClass().getName());
     }
 
     @Override
     public int hashCode() {
         // for data structures relying on hash first, all equal instances must have the same hash to avoid side effects
-        return getClass().hashCode();
+        return getClass().getName().hashCode();
     }
 }