+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/unique/support/SampleExt.java
@@ -1,0 +1,11 @@
+package ru.vyarus.dropwizard.guice.config.unique.support;
+
+import ru.vyarus.dropwizard.guice.module.installer.feature.eager.EagerSingleton;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 09.09.2019
+ */
+@EagerSingleton
+public class SampleExt {
+}