+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/unique/support/SampleModule.java
@@ -1,0 +1,10 @@
+package ru.vyarus.dropwizard.guice.config.unique.support;
+
+import ru.vyarus.dropwizard.guice.module.context.unique.item.UniqueModule;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 09.09.2019
+ */
+public class SampleModule extends UniqueModule {
+}