+++ CHANGELOG.md
@@ -100,6 +100,7 @@
         (but only if guice bindings analysis is not disabled)    
 * Fix configuration bindings for recursive configuration object declarations (#60)
 * Guicey version added into BOM (dependencyManagement section in guicey pom) to avoid duplicate versions declarations
+* Detect classes loaded by different class loader to correctly detect duplicate extensions
 
 Main breaking changes were caused by:
  - jersey 2.26 introduces an abstraction for injection layer in order to get rid of hk2 direct usage.